+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 FEATURES:
 
 - **New Resource**: `hcloud_rnds` Manage reverse DNS entries for servers and Floating IPs.
+* **New Resource**: `hcloud_floating_ip_assignment` Manage the association between Floating IPs and servers.
 - **New Datasource**: `hcloud_floating_ip` Lookup Floating ips.
 - **New Provider Config**: `poll_interval`  Configures the interval in which actions are polled by the client. Default `500ms`. Increase this interval if you run into rate limiting errors.